+++ R/helpers.R
@@ -34,7 +34,10 @@ lrn = function(.key, ...) {
   pkgs = suppressWarnings(mlr3::lrn(.key))$packages
   tryCatch(mlr3misc::require_namespaces(pkgs),
            packageNotFoundError = function(e) {
-             mlr3misc::stopf('Required packages not installed, please run `install_learners("%s")`.', .key) # nolint
+             mlr3misc::stopf(
+               "Required packages not installed, please run `install_learners('%s'`",
+               .key
+             ) # nolint
            })
 
   mlr3misc::dictionary_sugar_get(mlr_learners, .key, ...)
@@ -49,8 +52,6 @@ lrns = function(.keys, ...) {
   lapply(.keys, lrn, ...)
 }
 
-# copied from mlr3learners
-# pp - positive class predicted probabilities
 pprob_to_matrix <- function(pp, task) {
   y <- matrix(c(pp, 1 - pp), ncol = 2L, nrow = length(pp))
   colnames(y) <- task$class_names