+++ inst/paramtest/test_paramtest_LiblineaR_classif_liblinear.R
@@ -29,5 +29,6 @@ test_that("classif.liblinear predict", {
   ParamTest = run_paramtest(learner, fun, exclude)
   expect_true(ParamTest, info = paste0(
     "Missing parameters:",
-    paste0("- '", ParamTest$missing, "'", collapse = "")))
+    paste0("- '", ParamTest$missing, "'", collapse = "")
+  ))
 })