+++ tests/testthat/helper.R
@@ -5,7 +5,7 @@ lapply(list.files(system.file("testthat", package = "mlr3"),
   pattern = "^helper.*\\.[rR]", full.names = TRUE), source)
 
 load_tests = function(key) {
-  class = subset(list_mlr3learners(), id == key)$class
+  class = strsplit(mlr3::lrn(key)$id, ".", TRUE)[[1]][[1]]
 
   if (class %in% c("surv", "dens")) {
     if (!requireNamespace("mlr3proba", quietly = TRUE)) {
@@ -22,5 +22,4 @@ load_tests = function(key) {
                       pattern = "^helper.*\\.[rR]", full.names = TRUE), source))
     require("mlr3cluster")
   }
-
 }